backend-sous/requirements.txt
flask
sqlalchemy
psycopg2-binary
flask-sqlalchemy
python-dotenv